app/requirements.txt
# === FastAPI & Core ===
fastapi==0.124.2
uvicorn==0.30.6
gunicorn==23.0.0
starlette==0.50.0
python-multipart==0.0.9
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
email-validator==2.2.0
PyYAML==6.0.1
pytz==2024.1

# === Auth ===
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
PyJWT==2.10.1

# === HTTP & Utils ===
requests==2.32.3
httpx==0.28.1

# === Database ===
sqlalchemy==2.0.36
alembic==1.17.2
asyncpg==0.31.0
psycopg2-binary==2.9.11
greenlet==3.3.0

# Minio
minio

# === Extensions ===
fastapi-pagination==0.15.2
sentry-sdk==2.18.0
pillow==12.0.0

#celery
celery

redis==7.2.0
fastapi-cache2==0.2.2
motor==3.7.1

# === LLM ===
openai==2.13.0
